feat: Gets cytoscape ts code to run in an ipynb
Cytoscape JS code successfully runs in an ipynb.
!!NOTE!!: The ipynb graph renders as a very narrow
window. The fulfillment plan icons are also not
rendering, they were downloaded as PNGs which can't be
embedded directly to the JS build artifact afaik and
can't be accessed by the notebookjs generated HTML
file that renders the graph in the ipynb. If they're
SVGs it should work.