feat: ipynb renders the graph in a larger window
The ipynb now renders the graph in a larger window,
such that it's more visible. The maker background image
is also set to a URL of an SVG, which works nicely.
The tooltip code is refactored to a new file, and
the cytoscape elements, layout and style are now
represented as JSON files, read by the ipynb and passed
to the drawGraph function, so the graph is now fully
configurable by the python code.