feat: Makes things more dev-friendly
Adds a script, cmd/buildHTML.py, which monitors if the
JS is rebuilt, or if any of the elements.json, style.json
or layout.json files are changed. If they are, it rebuilds
the HTML using notebookjs, so you can basically make changes
to those files and see them immediately.